Support multiple transaction types
There are at least three transaction types supported in Ethereum, and
this tool was previously only supporting the legacy type.  This
refactoring makes this more explicit, and catches attempts to parse
other transaction types as legacy transactions.  At this stage, however,
I have not added support for the other transaction types.  That is to be
done in a separate issue.