`Transaction.Expectation` => `Transaction.Exception`
This resolves some issues around the notion of an "expectation" in the
tests.  The key is that this is really about failures at the level of a
transaction, rather than failures within the EVM itself.

As a result, I have renamed this accordingly and added some missing
cases.  I've also removed cases that aren't used and didn't correspond
to anything real.