Support Updated StateTest format
This updates the tool base to support the slightly extended StateTest
format.  For example, this now includes fields such as
`currentWithdrawalsRoot` and `currentRandom` which are needed
post-merge.

At this stage, there is an outstanding issue related to hashes however,
in that Geth is expecting some to be provided ... but they are not
present in the state test files themselves.  Potentially they need to be
generated by the tool?